requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
#discord==2.3.2
discord.py==2.4.0
frozenlist==1.4.0
greenlet==3.0.3
idna==3.4
#multidict==6.0.4
numpy==1.26.4
psycopg2-binary==2.9.9
py-cord==2.0.0b1
pygame==2.0.1
python-dotenv==1.0.1
requests==2.31.0
scipy==1.13.1
SQLAlchemy==2.0.32
table2ascii==1.1.3
typing_extensions==4.12.2
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.9.2
